requirements.txt
langchain-community==0.3.0
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
tqdm